shop card is done + tests

+++ services/shop_cards/server/db/models.go
@@ -8,7 +8,7 @@ type ShopCard struct {
 	CreatedAt   time.Time `gorm:"autoCreateTime"`
 	Description string    `gorm:"type:varchar(500)"`
 	Category    string    `gorm:"type:varchar(100)"`
-	Price       *float64  `gorm:"type:decimal(10,2);"`
+	Price       *float32  `gorm:"type:decimal(10,2);"`
 	Visible     bool      `gorm:"default:true"`
 	CoverURL    *string   `gorm:"type:varchar(100)"`
 }
